% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_by.R
\name{deduplicate_by}
\alias{deduplicate_by}
\title{Deduplicate a DataFrame.}
\usage{
deduplicate_by(.data, ...)
}
\arguments{
\item{.data}{DataFrame to be deduplicated}

\item{...}{columns to deduplicate by}
}
\value{
deduplicated DataFrame
}
\description{
A tidy function that deduplicates a DataFrame by unique values
in the user supplied columns. Keeps the first row for each value.
}
