% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_thematic_map.R
\name{make_thematic_map}
\alias{make_thematic_map}
\title{Make a thematic map.}
\usage{
make_thematic_map(.data, fill_var, .map = NULL, join_by = NULL,
  remove_na = TRUE, background_map = NULL, breaks = NULL,
  break_name_expr = "\\\\d+\\\\.?\\\\d+(?=\\\\])", scale_name = NULL)
}
\arguments{
\item{.data}{DataFrame or sf DataFrame with the data to plot.}

\item{fill_var}{column to map to the fill aesthetic.}

\item{.map}{sf DataFrame of the map to fill, if `.data`` is not already an sf object.}

\item{join_by}{column or dplyr join specification to join `.data`` and `.map by`.}

\item{remove_na}{whether to remove shapes that have NA values or not.}

\item{background_map}{sf DataFrame of a map to plot as the background.}

\item{breaks}{number of vector of breaks for the colour scale of `fill_var`.}

\item{break_name_expr}{regex to clean up break names with.}

\item{scale_name}{name for the fill scale label.}
}
\value{
ggplot thematic map
}
\description{
Makes a quick thematic (choropleth) map from data provided,
with a fixed colour scheme.
}
