% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg_level2.R
\docType{data}
\name{tdwg_level2}
\alias{tdwg_level2}
\title{World Geographical Scheme for Recording Plant Distributions, Level 2}
\format{An object of class \code{"sf"}.}
\source{
\href{https://github.com/tdwg/wgsrpd}{TDWG Github}
}
\usage{
data(tdwg_level2)
}
\description{
Simple feature collection of WGSRPD level 1 regions defined by
Biodiversity Information Standards (TDWG). Level 2 corresponds
to sub-continental regions.
}
\references{
Brummitt RK (2001) World Geographical Scheme for Recording Plant Distributions, Edition 2.
Biodiversity Information Standards (TDWG) \href{http://www.tdwg.org/standards/109}{http://www.tdwg.org/standards/109}
}
\keyword{datasets}
