% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg_level3.R
\docType{data}
\name{tdwg_level3}
\alias{tdwg_level3}
\title{World Geographical Scheme for Recording Plant Distributions, Level 3}
\format{An object of class \code{"sf"}.}
\source{
\href{https://github.com/tdwg/wgsrpd}{TDWG Github}
}
\usage{
data(tdwg_level3)
}
\description{
Simple feature collection of WGSRPD level 3 regions defined by
Biodiversity Information Standards (TDWG). Level 3 corresponds
to "Botanical Countries". These may ignore purely political considerations,
and where there are no complicating political factors may be the same
as Level 4 regions. Very large countries have been subdivided according to
states or provinces.
}
\references{
Brummitt RK (2001) World Geographical Scheme for Recording Plant Distributions, Edition 2.
Biodiversity Information Standards (TDWG) \href{http://www.tdwg.org/standards/109}{http://www.tdwg.org/standards/109}
}
\keyword{datasets}
