% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg_level4.R
\docType{data}
\name{tdwg_level4}
\alias{tdwg_level4}
\title{World Geographical Scheme for Recording Plant Distributions, Level 4}
\format{An object of class \code{"sf"}.}
\source{
\href{https://github.com/tdwg/wgsrpd}{TDWG Github}
}
\usage{
data(tdwg_level4)
}
\description{
Simple feature collection of WGSRPD level 4 regions defined by
Biodiversity Information Standards (TDWG). Level 4 corresponds
to "Basic Recording Units". These fully recognise political integrity,
but where there are no complicating political factors may be the same
as Level 3 regions.
}
\references{
Brummitt RK (2001) World Geographical Scheme for Recording Plant Distributions, Edition 2.
Biodiversity Information Standards (TDWG) \href{http://www.tdwg.org/standards/109}{http://www.tdwg.org/standards/109}
}
\keyword{datasets}
